NA-pad blink events / missing data

Deblinking (a.k.a. NA-padding) of time series data. The intended use of this method is to remove blink-related artifacts surrounding periods of missing data. For instance, when an individual blinks, there are usually rapid decreases followed by increases in pupil size, with a chunk of data missing in-between these 'spike'-looking events. The deblinking procedure here will NA-pad each missing data point by your specified number of ms.

Usage

deblink(eyeris, extend = 40)

Arguments

eyeris

An object of class eyeris dervived from load().

extend

Either a single number indicating the number of milliseconds to pad forward/backward around each missing sample, or, a vector of length two indicating different numbers of milliseconds pad forward/backward around each missing sample, in the format c(backward, forward).

Value

An eyeris object with a new column: pupil_deblink.

Examples

if (FALSE) { # \dontrun{
system.file("extdata", "assocret.asc", package = "eyeris") |>
  eyeris::load_asc() |>
  eyeris::deblink(extend = 40) # 40 ms in both directions
} # }

system.file("extdata", "assocret.asc", package = "eyeris") |>
  eyeris::load_asc() |>
  eyeris::deblink(extend = c(40, 50)) # 40 ms backward, 50 ms forward
#> $file
#> [1] "/opt/homebrew/lib/R/4.4/site-library/eyeris/extdata/assocret.asc"
#> 
#> $timeseries
#> # A tibble: 22,145 × 8
#>    time_orig eye_x eye_y eye      hz type     pupil_raw pupil_raw_deblink
#>        <int> <dbl> <dbl> <chr> <dbl> <chr>        <dbl>             <dbl>
#>  1  11242177  949.  646. R      1000 diameter      6174              6174
#>  2  11242178  947.  644. R      1000 diameter      6176              6176
#>  3  11242179  946.  643  R      1000 diameter      6178              6178
#>  4  11242180  945.  642  R      1000 diameter      6179              6179
#>  5  11242181  945.  641. R      1000 diameter      6180              6180
#>  6  11242182  945.  642. R      1000 diameter      6182              6182
#>  7  11242183  945.  644. R      1000 diameter      6187              6187
#>  8  11242184  945.  644. R      1000 diameter      6197              6197
#>  9  11242185  945.  643. R      1000 diameter      6207              6207
#> 10  11242186  945.  642. R      1000 diameter      6203              6203
#> # ℹ 22,135 more rows
#> 
#> $events
#> # A tibble: 67 × 3
#>    block     time text                      
#>    <dbl>    <dbl> <chr>                     
#>  1     1 11242177 !MODE RECORD CR 1000 2 1 R
#>  2     1 11243349 EXP_START                 
#>  3     1 11243355 TRIALID 0                 
#>  4     1 11243356 START_TRIAL_0             
#>  5     1 11243356 CUE_START_0               
#>  6     1 11244328 CUE_STOP_0                
#>  7     1 11244329 DELAY_START_0             
#>  8     1 11245337 DELAY_STOP_0              
#>  9     1 11245338 PROBE_START_0             
#> 10     1 11245956 RESPONSE_0                
#> # ℹ 57 more rows
#> 
#> $blinks
#> # A tibble: 0 × 5
#> # ℹ 5 variables: block <int>, stime <dbl>, etime <dbl>, dur <dbl>, eye <fct>
#> 
#> $info
#>                  date        model version sample.rate   cr  left right mono
#> 1 2009-01-11 02:40:47 EyeLink 1000   4.594        1000 TRUE FALSE  TRUE TRUE
#>   screen.x screen.y                                 mount filter.level
#> 1     1920     1080 Desktop / Monocular / Head Stabilized            2
#>   sample.dtype event.dtype pupil.dtype velocity resolution htarg input buttons
#> 1         GAZE        GAZE    DIAMETER    FALSE      FALSE FALSE FALSE   FALSE
#> 
#> $latest
#> [1] "pupil_raw_deblink"
#> 
#> attr(,"class")
#> [1] "eyeris"